Start to version our testing libraries.
This change has two main changes
1) Move the functional options under `./pkg/testing/v1alpha1` (this also splits out `{service,route,configuration,revision}.go` with the options for this types)
2) Move the test of `./pkg/reconciler/testing` under a `v1alpha1` subdirectory (since it all deals almost exclusively with `v1alpha1` stuff).

Part of the motivation for this separation is that the functional options are used in the e2e tests, which are unrelated to `./pkg/reconciler`.
