+++ pkg/reconciler/serverlessservice/serverlessservice_test.go
@@ -47,7 +47,8 @@ import (
 	clientgotesting "k8s.io/client-go/testing"
 
 	. "github.com/knative/pkg/reconciler/testing"
-	. "github.com/knative/serving/pkg/reconciler/testing"
+	. "github.com/knative/serving/pkg/reconciler/testing/v1alpha1"
+	. "github.com/knative/serving/pkg/testing"
 )
 
 func TestNewController(t *testing.T) {
