+++ pkg/testing/OWNERS
@@ -1,0 +1,12 @@
+# The OWNERS file is used by prow to automatically merge approved PRs.
+
+approvers:
+- productivity-approvers
+- serving-api-approvers
+
+reviewers:
+- productivity-reviewers
+- serving-api-approvers
+
+labels:
+- area/test-and-release
