+++ test/conformance/runtime/util.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	pkgTest "github.com/knative/pkg/test"
-	reconciler "github.com/knative/serving/pkg/reconciler/testing"
+	reconciler "github.com/knative/serving/pkg/testing/v1alpha1"
 
 	"github.com/knative/serving/test"
 	"github.com/knative/serving/test/types"
