+++ test/e2e/image_pull_error_test.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
 	serviceresourcenames "github.com/knative/serving/pkg/reconciler/service/resources/names"
-	v1alpha1testing "github.com/knative/serving/pkg/reconciler/testing"
+	v1alpha1testing "github.com/knative/serving/pkg/testing/v1alpha1"
 	"github.com/knative/serving/test"
 )
