+++ test/performance/scale_revision_by_load_test.go
@@ -28,8 +28,8 @@ import (
 	"github.com/knative/pkg/controller"
 	pkgTest "github.com/knative/pkg/test"
 	ingress "github.com/knative/pkg/test/ingress"
-	testingv1alpha1 "github.com/knative/serving/pkg/reconciler/testing"
 	"github.com/knative/serving/pkg/resources"
+	testingv1alpha1 "github.com/knative/serving/pkg/testing/v1alpha1"
 	"github.com/knative/serving/test"
 	"github.com/knative/test-infra/shared/junit"
 	"github.com/knative/test-infra/shared/loadgenerator"
